fix: correct foreign key type mismatches in migrations for MariaDB/MySQL
The `users.id` column is `int unsigned` (not bigint) in this codebase.
Three migrations used `unsignedBigInteger` for FKs referencing `users.id`,
causing errno 150 FK constraint errors on fresh install:

- google_calendar_tokens.user_id: unsignedBigInteger → unsignedInteger
- conversations.created_by: unsignedBigInteger → unsignedInteger
- conversation_participants.user_id: unsignedBigInteger → unsignedInteger
- messages.user_id: unsignedBigInteger → unsignedInteger

Also fix: remove ->after('timeline_group_id') in add_google_event_id
migration since that column was never created (migration stub was empty).

+++ database/migrations/2026_02_27_000001_create_google_calendar_tokens_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('google_calendar_tokens', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id')->unique();
+            $table->unsignedInteger('user_id')->unique();
             $table->text('access_token');
             $table->text('refresh_token')->nullable();
             $table->timestamp('expires_at')->nullable();
